renaming and changing location
