+++ config.json
@@ -1,6 +1,6 @@
 {
-    "width": 1920,
-    "height": 1080,
+    "width": 1280,
+    "height": 720,
 
     "fullscreen" : false,
     "landscape" : true,
